+++ src/app/modules/email/components/email-editor/email-editor.component.ts
@@ -9,7 +9,7 @@ import { Email } from '../../classes/email/email';
 import { EmailContent } from '../../classes/email-content/email-content';
 import { EmailService } from '../../services/email/email.service';
 import { PgpService } from '../../services/pgp/pgp.service';
-import { Oidc2AttachmentService } from '../../services/oidc2-attachment/oidc2-attachment.service';
+import { PgpKeyCertificationService } from '../../services/pgp-key-certification/pgp-key-certification.service';
 
 @Component({
   selector: 'app-email-editor',
@@ -34,7 +34,7 @@ export class EmailEditorComponent implements OnInit {
     private readonly emailService: EmailService,
     private readonly pgpService: PgpService,
     private readonly identityService: IdentityService,
-    private readonly oidc2AttachmentService: Oidc2AttachmentService,
+    private readonly oidc2AttachmentService: PgpKeyCertificationService,
   ) { 
     this.emailForm.controls.from.valueChanges.subscribe(() => this.updateEncryptionDisabledState());
     this.emailForm.controls.to.valueChanges.subscribe(() => this.updateEncryptionDisabledState());
